+++ packages/auth/src/index.ts
@@ -2,6 +2,7 @@ import 'reflect-metadata';
 
 export * from './constants/index.js';
 export * from './interfaces/index.js';
+export type { AuthConfig, AuthUser } from './interfaces/index.js';
 export * from './services/jwt.service.js';
 export * from './services/grpc-metadata.helper.js';
 export * from './auth.module.js';